API/requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
django-cors-headers==3.6.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==2.0.3
openapi-codec==1.3.2
packaging==20.8
psycopg2==2.8.6
pylint==2.6.0
pyparsing==2.4.7
pytz==2020.5
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
uritemplate==3.0.1
urllib3==1.26.2
wrapt==1.12.1
